sql_config/requirements.txt
# requirements.txt
pydantic
mysql-connector-python
python-dotenv
